Fix: Peripheral's listenValues() API now buffers items
This is an issue we encountered when developing iOSOtaLibrary, which is that if we added the network request part after receiving a new Characteristic value, then we'd only get the one. If the Memfault device sent 8-9-11 chunks, we'd only get one. Removing the line of code to "await network request" made all chunks show up and print on the Terminal. The issue was the upstream PassthroughSubject dropping the items, because it doesn't buffer items. Unlike an AsyncStreamSequence.