Minor code cleanup to Peripheral read/write Descriptor functions
They looked to be implemented in a very similar way to the code that was causing the issue when reading a Characteristic, so I decided to clean this part up as well. But, this code did look like it was working. Hopefully we're not doing any further harm with these changes.